// management/src/routes/roles/__generated__/RolesManagerFragment.graphql.ts
export type RolesManagerFragment$data = {
  readonly roles: {
    readonly edges: ReadonlyArray<{
      readonly cursor: string;
      readonly node: {
        readonly capabilities: any | null | undefined;
        readonly createdAt: any;
        readonly description: string | null | undefined;
        readonly id: string;
        readonly name: string;
        readonly slug: string;
        readonly updatedAt: any;
      } | null | undefined;
    }>;
    readonly pageInfo: {
      readonly endCursor: string | null | undefined;
      readonly hasNextPage: boolean;
    };
    readonly totalCount: number;
  };
  readonly " $fragmentType": "RolesManagerFragment";
};
export type RolesManagerFragment$key = {
  readonly " $data"?: RolesManagerFragment$data;
  readonly " $fragmentSpreads": FragmentRefs<"RolesManagerFragment">;
};